+++ lucene/suggest/src/test/org/apache/lucene/search/suggest/LookupBenchmarkTest.java
@@ -112,7 +112,7 @@ public class LookupBenchmarkTest extends LuceneTestCase {
     while ((line = br.readLine()) != null) {
       int tab = line.indexOf('|');
       assertTrue("No | separator?: " + line, tab >= 0);
-      int weight = Integer.parseInt(line.substring(tab + 1));
+      long weight = Long.parseLong(line.substring(tab + 1));
       String key = line.substring(0, tab);
       input.add(new Input(key, weight));
     }
